Split ExtensionManager class out of ExtensionService for use outside the engine
